auth_api/python/requirements.txt
Flask
PyJWT
mysql-connector-python
python-dotenv